+++ src/main/java/dansplugins/recipesystem/utils/RecipeRegistry.java
@@ -1,7 +1,7 @@
 package dansplugins.recipesystem.utils;
 
 import dansplugins.recipesystem.MoreRecipes;
-import dansplugins.recipesystem.objects.*;
+import dansplugins.recipesystem.objects.MoreRecipesItem;
 import dansplugins.recipesystem.services.ItemStackService;
 
 public class RecipeRegistry {
@@ -13,60 +13,13 @@ public class RecipeRegistry {
         this.moreRecipes = moreRecipes;
     }
 
+    /**
+     * Registers the crafting recipe of every item in the catalog.
+     */
     public void registerRecipes() {
-        Saddle saddle = new Saddle(itemStackService, moreRecipes);
-        saddle.registerRecipe();
-
-        NameTag nametag = new NameTag(itemStackService, moreRecipes);
-        nametag.registerRecipe();
-
-        TotemOfUndying totemOfUndying = new TotemOfUndying(itemStackService, moreRecipes);
-        totemOfUndying.registerRecipe();
-
-        GrassBlock grassBlock = new GrassBlock(itemStackService, moreRecipes);
-        grassBlock.registerRecipe();
-
-        StringItem string = new StringItem(itemStackService, moreRecipes);
-        string.registerRecipe();
-
-        Lead lead = new Lead(itemStackService, moreRecipes);
-        lead.registerRecipe();
-
-        BlazeRod blazeRod = new BlazeRod(itemStackService, moreRecipes);
-        blazeRod.registerRecipe();
-
-        ChainmailHelmet chainmailHelmet = new ChainmailHelmet(itemStackService, moreRecipes);
-        chainmailHelmet.registerRecipe();
-
-        ChainmailChestplate chainmailChestplate = new ChainmailChestplate(itemStackService, moreRecipes);
-        chainmailChestplate.registerRecipe();
-
-        ChainmailLeggings chainmailLeggings = new ChainmailLeggings(itemStackService, moreRecipes);
-        chainmailLeggings.registerRecipe();
-
-        ChainmailBoots chainmailBoots = new ChainmailBoots(itemStackService, moreRecipes);
-        chainmailBoots.registerRecipe();
-
-        IronHorseArmor ironHorseArmor = new IronHorseArmor(itemStackService, moreRecipes);
-        ironHorseArmor.registerRecipe();
-
-        GoldenHorseArmor goldenHorseArmor = new GoldenHorseArmor(itemStackService, moreRecipes);
-        goldenHorseArmor.registerRecipe();
-
-        DiamondHorseArmor diamondHorseArmor = new DiamondHorseArmor(itemStackService, moreRecipes);
-        diamondHorseArmor.registerRecipe();
-
-        Gunpowder gunpowder = new Gunpowder(itemStackService, moreRecipes);
-        gunpowder.registerRecipe();
-
-        Cobweb cobweb = new Cobweb(itemStackService, moreRecipes);
-        cobweb.registerRecipe();
-
-        SlimeBall slimeBall = new SlimeBall(itemStackService, moreRecipes);
-        slimeBall.registerRecipe();
-
-        PrismarineShard prismarineShard = new PrismarineShard(itemStackService, moreRecipes);
-        prismarineShard.registerRecipe();
+        for (MoreRecipesItem item : MoreRecipesItem.values()) {
+            item.registerRecipe(itemStackService, moreRecipes);
+        }
     }
 
 }